0009216: When editing a contract again, the template shows a date
When editing a contract again, the template shows a date

 - also use float insted of integer when handling daily working hours

https://forge.tine20.org/mantisbt/view.php?id=9216

Reviewed-on: https://gerrit.tine20.org/tine20/2503
Tested-by: Philipp Schüle <p.schuele@metaways.de>

+++ tine20/HumanResources/js/HumanResources.js
@@ -20,7 +20,7 @@ Tine.widgets.grid.RendererManager.register('HumanResources', 'Contract', 'workin
     var object = Ext.decode(v);
     var sum = 0;
     for (var i=0; i < object.days.length; i++) {
-        sum = sum + parseInt(object.days[i]);
+        sum = sum + parseFloat(object.days[i]);
     }
     return sum;
 });
